pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fidelfolio_dl"
version = "0.1.0"
description = "Deep Learning for Financial Valuation - FidelFolio Challenge"
readme = "README.md"
authors = [
  { name = "FidelFolio Team", email = "team@fidelfolio.com" },
]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "scikit-learn>=1.0.0",
    "tensorflow>=2.6.0",
    "pyyaml>=6.0",
]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "jupyterlab",
    "matplotlib",
    "seaborn"
]

[project.scripts]
fidelfolio-train = "main:main"
